+++ example/Direct_connection_problem/module/connection.js
@@ -8,20 +8,5 @@ const connection = new qmy(
     port: 3306,
     user: "<user>",
   },
-  //optional
-  {
-    settings: {
-      logfile: {
-        //This option saves errors to a file
-        status: true, //{default} false
-        path: "../", //{default} ./
-      },
-      connection: {
-        //{false} You will not see connection message in console log
-        //{true} You will see connection message in console log
-        log: false, //{default} true
-      },
-    },
-  }
 );
 module.exports = connection;