+++ index.js
@@ -4,8 +4,6 @@
 const mysql = require("mysql");
 const got = require("got");
 const v = require("./package.json");
-const config = require("./src/configs/settings.json");
-const chalk = require("chalk");
 //METHODS
 const get = require("./src/methods/get");
 const set = require("./src/methods/set");
@@ -13,74 +11,64 @@ const update = require("./src/methods/update");
 const remove = require("./src/methods/remove");
 const query = require("./src/methods/query");
 const custom = require("./src/methods/log/custom");
-const check = require("./src/methods/other/check");
-const output = require("./src/methods/log/output");
-(async () => {
-  try {
-    const data = await got.post("https://img.shields.io/npm/v/qmy.json", {
-      responseType: "json",
-    });
-    if ("v" + v.version !== `${data.body.value}`) {
-      console.log(
-        `
-╔═════════════╦═════════════════════════════════╗
-║ New Version ║ ${data.body.value}                          ║
-╠═════════════╬═════════════════════════════════╣
-║ Install     ║ npm install qmy                 ║
-╠═════════════╬═════════════════════════════════╣
-║ #Note       ║ When you update to the latest   ║
-║             ║ version, you don't get any bugs ║
-╚═════════════╩═════════════════════════════════╝
-      `
-      );
-    }
-  } catch (error) {}
-})();
+const parseUrl = require("./src/methods/other/parseUrl");
+/**
+ * Create a new Connection 
+ ```js
+ (connectionUri: string | ConnectionConfig): mysql.Connection
+ ```
+ */
 function connection(
-  {
-    user: user,
-    host: host,
-    password: password,
-    database: database,
-    port: port,
+  ConnectionConfig = {
+    host: "localhost",
+    port: 3306,
+    localAddress: undefined,
+    socketPath: undefined,
+    user: undefined,
+    password: undefined,
+    database: undefined,
+    connectTimeout: 10000,
+    insecureAuth: false,
+    supportBigNumbers: false,
+    bigNumberStrings: false,
+    dateStrings: false,
+    debug: undefined,
+    trace: true,
+    stringifyObjects: false,
+    timezone: "local",
+    flags: "",
+    queryFormat: undefined,
+    pool: undefined,
+    ssl: false,
+    localInfile: true,
+    multipleStatements: false,
+    typeCast: true,
+    maxPacketSize: 0,
+    charsetNumber: 33,
+    clientFlags: 455631,
+    protocol41: true,
   },
   options = {
-    settings: {
-      logfile: {
-        status: (statuslog = false),
-        path: (pathlog = "./"),
-      },
-      connection: {
-        log: (connectionlog = true),
-      },
+    connection: {
+      log: (connectionlog = true),
     },
   }
 ) {
   try {
-    var connectionlog = options.settings.connection.log;
-    var statuslog = options.settings.logfile.status;
-    var pathlog = options.settings.logfile.path;
+    var database = ConnectionConfig.database;
+    if (typeof ConnectionConfig === "string") {
+      database = parseUrl(ConnectionConfig).database;
+    }
+    var connectionlog = options.connection.log;
   } catch {}
-  connectionlog = check(connectionlog, true);
-  statuslog = check(statuslog, false);
-  pathlog = check(pathlog, "./");
   const createCon = () => {
-    let db = mysql.createConnection({
-      user: user,
-      host: host,
-      password: password,
-      database: database,
-      port: port,
-    });
+    let db = mysql.createConnection(ConnectionConfig);
     return db;
   };
-  function out(err) {
-    return output(err, statuslog, pathlog);
-  }
   const reConnect = (connection, error) =>
     new Promise((resolve) => {
       if (!error.toString().includes("Packets out of order")) {
-        return out(error);
+        throw error;
       }
       connection.destroy();
       if (connectionlog == true)
@@ -94,10 +82,10 @@ function connection(
       con.connect(async (err) => {
         if (err) {
           if (!err.code.toString().includes("Packets out of order")) {
-            return out(err.code);
+            throw err;
           }
           try {
-            throw new Error(err.code);
+            throw err;
           } catch (e) {
             await reConnect(con, e);
             Connect();
@@ -125,94 +113,150 @@ function connection(
   // <-  Data get ->
   /**
 *This function is used to get data from databases
-*
-* @example get("<table>", "<Where>:<PrimaryKey>", function (result) {
-  if (!result[0]) {
-    //false
+* @param  {array: <default:false>:boolean }
+* @example get("<table>", "<Where>:<PrimaryKey>").then((result) => {
+    console.log(result);
+  }).catch((err) => {
+    console.log(err);
+  });
+* Example of a <Where>:<PrimaryKey>
+ @example get("users", "id:1111111111").then((result) => {
     console.log(false);
-  } else if (result[0]) {
-    //true
-    console.log(result[0]);
-  }
-});
+  }).catch((err) => {
+    console.log(err);
+  });
 */
-  this.get = function (table, PrimaryKey, callback) {
-    return get(
-      table,
-      PrimaryKey,
-      callback,
-      statuslog,
-      pathlog,
-      Connection,
-      database
-    );
+  this.get = function (table, PrimaryKey, options = { array: false }) {
+    return get(table, PrimaryKey, Connection, database, options);
   };
   // <- query ->
   /**
 *This is a query function that you can perform any operation with this function. 
 *This function is useful for developers who want to perform an operation that is not available in existing functions
-* @example query("<sql>",function(result) {
+* @param  {array: <default:true>:boolean }
+* @example query("<sql>",{ array: false }).then((result) => {
     console.log(result)
-})
+}).catch((err) => {
+    console.log(err);
+  });
 */
-  this.query = function (sql, callback) {
-    return query(sql, callback, statuslog, pathlog, Connection);
+  this.query = function (sql, options = { array: true }) {
+    return query(sql, Connection, options);
   };
   // <- Data remove ->
   /**
-   *This function is used to remove data
-   *
-   * @example remove("<table>", "<Where>:<PrimaryKey>")
+  *This function is used to remove data
+  *
+  * @example remove("<table>", "<Where>:<PrimaryKey>").then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
+  * Example of a <Where>:<PrimaryKey>
+ @example 
+  remove("users", "id:1111111111").then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
    */
-  this.remove = function (table, PrimaryKey, callback) {
-    return remove(
-      table,
-      PrimaryKey,
-      callback,
-      statuslog,
-      pathlog,
-      Connection,
-      database
-    );
+  this.remove = function (table, PrimaryKey) {
+    return remove(table, PrimaryKey, Connection, database, options);
   };
   // <- Data storage ->
   /**
    *This function is used to save data in databases
-   *
-   * @example set({ "table": "<table>", "column": "<column>", "values": "<values>"})//It is placed "," if there is more than one value or column
+   * @param  {sign: <default:true>:boolean }
+   * @example set({
+    "table": "<table>", 
+    "column": "<column>",
+     "values": "<values>"
+    }).then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
+   //It is placed sign if there is more than one value or column
+   @example set({
+    "table": "users", 
+    "column": "id - arth - color",
+    "values": "1111111111 - arth - red"}, { sign:"-" }).then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
    */
   this.set = function (
     { table: table, column: column, values: values },
-    callback
+    options = { sign: "," }
   ) {
     return set(
       { table: table, column: column, values: values },
-      callback,
-      statuslog,
-      pathlog,
       Connection,
-      database
+      database,
+      options
     );
   };
   // <- Data update ->
   /**
    *This function is used to update data in databases
    *
-   * @example update({ table: "<table>", column: "<column>", PrimaryKey: "<Where>:<PrimaryKey>", value: "<value>" })
+   * @example update({
+       table: "<table>",
+       column: "<column>",
+       PrimaryKey: "<Where>:<PrimaryKey>",
+       value: "<value>",
+     }).then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
+   * Example of a <Where>:<PrimaryKey>
+ @example 
+ update({
+       table: "users",
+       column: "name",
+       PrimaryKey: "id:1111111111",
+       value: "arth",
+     }).then((result) => {
+    console.log(result)
+ }).catch((err) => {
+    console.log(err);
+  });
    */
-  this.update = function (
-    { table: table, column: column, value: value, PrimaryKey: PrimaryKey },
-    callback
-  ) {
+  this.update = function ({
+    table: table,
+    column: column,
+    value: value,
+    PrimaryKey: PrimaryKey,
+  }) {
     return update(
       { table: table, column: column, value: value, PrimaryKey: PrimaryKey },
-      callback,
-      statuslog,
-      pathlog,
       Connection,
-      database
+      database,
+      options
     );
   };
+  (async () => {
+    try {
+      const data = await got.post("https://img.shields.io/npm/v/qmy.json", {
+        responseType: "json",
+      });
+      if ("v" + v.version !== `${data.body.value}`) {
+        console.log(
+          `
+# New Version - ${data.body.value}           
+╔═════════════╦═════════════════════════════════╗              
+║ Install     ║ npm install qmy                 ║
+╠═════════════╬═════════════════════════════════╣
+║ #Note       ║ When you update to the latest   ║
+║             ║ version, you don't get any bugs ║
+╚═════════════╩═════════════════════════════════╝
+      `
+        );
+      }
+    } catch (error) {}
+  })();
   //<- END ->
 }
 module.exports = connection;