+++ src/methods/update.js
@@ -2,38 +2,14 @@
  * @Copyright 2021 Arth(https://github.com/4i8/)
  */
 const sqlprocessor = require("./other/sqlprocessor");
-const warn = require("./log/warn");
-const output = require("./log/output");
-const e = require("../configs/error.json");
 function update(
   { table: table, column: column, value: value, PrimaryKey: PrimaryKey },
-  callback,
-  statuslog,
-  pathlog,
   Connection,
   database
 ) {
-  const pkey = PrimaryKey.split(":");
-  function out(err) {
-    return output(err, statuslog, pathlog);
-  }
-  if (
-    !table ||
-    !column ||
-    typeof value == "undefined" ||
-    !PrimaryKey ||
-    pkey.length > 2 ||
-    pkey.length <= 1
-  ) {
-    return warn(
-      "update",
-      `
-      update({ table: "<table>", column: "<column>", PrimaryKey: "<Where>:<PrimaryKey>", value: "<value>" })
-      `
-    );
-  }
-  //START
-  try {
+  return new Promise((resolve, reject) => {
+    //START
+    const pkey = PrimaryKey.split(":");
     Connection.query(
       `UPDATE ${sqlprocessor(database, "`")}.${sqlprocessor(
         table,
@@ -41,24 +17,12 @@ function update(
       )} SET \`${column}\` = '${value}' WHERE \`${pkey[0]}\` = '${pkey[1]}'`,
       function (err, result) {
         if (err) {
-          return out(err);
+          return reject(err);
         }
-        if (callback) {
-          callback(result);
-        }
+        resolve(result);
       }
     );
-  } catch (erroR) {
-    if (erroR.toString() === e.query_undefined) {
-      out(
-        `Solve this problem in this link:
-https://github.com/4i8/qmy/tree/main/example/Direct_connection_problem\n` +
-          e.query_undefined
-      );
-    } else {
-      out(erroR);
-    }
-  }
+  });
   //END
 }
 module.exports = update;